Clean up GovernanceProcess to reflect reality
In the DefCore Scale cycle, we moved some older process documents
out of the OpenStack wiki and into Gerrit in order to more easily
maintain them.  The original imports were largely a "lift and shift"
of the wiki text, with improvements and cleanups to be driven later.

The Governance Process document in particular is quite dated and
contains information that has changed substantially over time.  This
patch attempts to clean it up by modfiying it to reflect current
reality while also maintaing some flexibility for the future.  For
example, we haven't used Google Hangouts for meetings in many months,
so the patch simply states that meetings will be open and the means
to join them will be posted on the wiki and/or mailing list (thus
allowing us to use other collaboration tools in the future without
needing to ammend process).