Add links to .json in .rst files
The .rst Guideline files currently say that they're generated from
"the master JSON verison" but don't actually contain a link to their
JSON counterparts.  This patch makes that reference a hyperlink so
folks can more easily locate the JSON files that are the authoritative
documents.  This is especially useful when viewing rendered versions
of the RST docs.

Change-Id: Idbdd7d6825aa8af2404b9e2791fa67240cdef239