Add script to tabulate scores
During last week's DefCore sprint in Austin, we decided to move
to a plaintext file format for our scoring working document.  Some
participants expressed a desire to have a script to automatically
tabulate total scores for capabilities scored in the new simple
text format, and also to be able to export them to a CSV file
for those who prefer to visualize the data with spreadsheet
software.

This patch adds a utility script to do just that.  It reads a DefCore
Guideline JSON file to collect Criteria names and weights.  It also
reads a scoring file in the new simple text format to pull individual
Criteria scores for each Capability listed.  Using the weights and
Criteria scores for each capability, it then computes a total score
for the Capability.  It creates two output files: a text document
that mirrors the scoring file but with an additional column for the
total score, and a CSV file complete with column headers.  The total
scores are also printed to stdout.

The source and destination files are configurable via command line
arguments (with reasonable defaults so the script can be run with
no arguments at all).  The script also includes help output that
can be printed by running it with the standard -h or --help arguments.

Change-Id: I7539728f1e907bebc28609977251c9f580400499