[RFC] Remove Advisory Status
When adhearance to the DefCore Guidelines became mandatory in order
to acquire the OpenStack Powered logo, a provision was in place in the
DefCore Committee's processes that stated that new Capabilities and
Designated Sections would pass through a 6-month (or greater) "Advisory"
phase before becoming required.  The central idea behind this was to
provide a period for the Committee to receive feedback from the
community about whether the Capabilties/Designated Sections were valid
candidates for inclusion, or whether there were factors the Committee
didn't take into account when adding those items to a Guideline.

The negative side effect of the advisory cycle is that it prevents
capabilities that actually meet all other Criteria across the releases
covered by a Guideline from becoming part of our interoperability
standard.  Further, in actual practice the DefCore Committee has
received very little feedback on Capabilities and Designated Sections
marked as advisory.  This may be partly due to the fact that DefCore has
been on an abbreviated schedule (publishing four Board-approved Guidelines
in five months in 2015 rather than one every six months), so feedback
cycles were fairly short.  However, there is also a case to made that
the community (vendors, users, etc) simply aren't paying much attention
to advisory items.  This theory may be backed up by the fact that
refstack-client wasn't always required and refstack.net didn't always
specifically show results of advisory tests (indeed, refstack.net didn't
always exist during this period).  Similarly, past instructions for how
to submit results for logo programs provided a way to run only required
tests, and conceivably many vendors chose to do so.

Where we have generally seen feedback in practice is once a Capability
becomes required, vendors provide feedback in the form of flag requests
that show valid reasons why a Capability shouldn't have been required.
Such requests often include information that would likely have changed
the Committee's scoring of a Capability, such as the Capability being
unsupported by all backends [1][2] or a Capability being poorly supported
by other parts of OpenStack to the point where vendors and operators
have a valid case for not deploying it [3].

In practice, these factors mean that the advisory cycle isn't
accomplishing it's goal of providing a feedback window--in fact, in
practice it's actually delaying feedback for 6 months or more.  We could
try to solve this through better eduction, but perhaps the more
pragmatic path is simply to adapt DefCore to the behavior of it's
audience rather than trying to force the reverse (bearing in mind, after
all, that DefCore has been communicating this process for over a year now
and it apparently hasn't caught on).  In practice, that means
accepting that flags on required capabilities and tests is the primary
mechanism in which the DefCore Committee receives feedback, and getting rid of
the advisory cycle so we can start the feedback loop sooner.

It also bears mentioning that even if the mandatory advisory cycle
didn't exist, we still have a multi-month feedback period between when
a Capability is proposed and when the Board approves it in a new
Guideline.  This period is inherent in the fact that DefCore generally
has a "solid draft" at the bi-annual design summit which includes newly
scored items [4] and the month of vendor "self-testing" that follows it
[5].  The Board doesn't vote on a new Guideline until a full three months
after the Summit [6], so there's still quite ample opportunity for vendors
to provide feedback on new inclusions even before raising a flag becomes
their only method of recourse.

This patch removes the advisory cycle from our processes.

[1] https://review.openstack.org/#/c/178895/
[2] https://review.openstack.org/215263/
[3] https://review.openstack.org/#/c/196154/
[4]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/process/2015A.rst#n22
[5]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/process/2015A.rst#n26
[6]
http://git.openstack.org/cgit/openstack/defcore/tree/doc/source/process/2015A.rst#n30

Change-Id: Iac63858c71008ef850fbd5b1a1ef62c10c12c45d