[RFC] Add TC and UC Representation to DefCore
In the time since adherance to DefCore Guidelines became mandatory for
vendors to apply for OpenStack Powered trademark programs, the DefCore
Committee has begun to bring interoperability challenges posed by
OpenStack's technical makeup to light.  Some of these challenges
originate with cross-project interaction.  Examples may include but are
not limited to:

* Some projects adopt major-version API changes at different paces.
  For example, Nova today uses Glance's v1 API although v2 has been
  designated as CURRENT and v3 is already in the works.  Neutronclient
  did not support the Keystone v3 API until late 2014 [1] and Horizon does
  not support domain-scoped tokens even in Kilo.[2]  Other projects
  began supporting the v3 API sooner, resulting in confusion about when
  users should be able to expect to use v3 across all components of
  their clouds.

* Projects sometimes have overlapping capabilities.  For example,
  OpenStack has long had two networking options (nova-net and Neutron).

* Projects are sometimes not aware of how changes in other projects may
  affect them or their users.  For example, some projects have expressed
  that they're unsure of the quality or completeness of new
  capabilities in other projects, and therefore are slow to adopt them.

Several of these sorts of issues fall squarely under the charter of the
Technical Committee [3] in that they deal with Commonality, Integration,
and Quality.  However, today there is no established, regular mechanism
for the DefCore Committee and Technical Committee to interact.  This is
not to say that communication hasn't occurred (indeed, some TC members
have occasionally participated in DefCore reviews, meetings, and mailing
list discussions).

Rather, it has become clear since DefCore Guideline
adherance became a requirement of logo programs that enough issues are
being brought to light that a more regular, prioritized communication
between the DefCore Committee (as a whole) and the Technical Committee
(as a whole) would be beneficial for the community as a whole.  A more
established flow of information between the two bodies would raise
awareness among the Technical Committee of issues being uncovered, and
would help the DefCore Committee make better decisions about
Capabilities to include and guidance to provide.  A regular report would
help with prioritization, and more frequent interaction between the
bodies would provide continuing discussion.

Similarly, DefCore Guidelines and the interoperability challenges that
the DefCore Committee brings to light ultimate impact the user
community.  In that capacity, the User Committee is chartered to serve
as point of information sharing and user feedback. [4]
A more established flow of information between the User
Committee and DefCore Committee would enable the User Committee to
provide input to DefCore decisions and enable stronger communication
channels to the end user community.

This patch adds ammendments to the DefCore Committee's procedures
recognizing representation on the DefCore Committee from members of the
Technical Committee and User Committee.  While the DefCore Committee
cannot mandate that members of other goverance bodies attend it's
meetings or participate in it's work, this patch serves to recognize
participation from those bodies.  Further, this patch establishes liaisons
from the DefCore Committee to the Technical and User Committees and mandates
regular participation in their meetings and discussions.  Separate
proposals will be made to the User Committee and Technical Committee
that they adopt reciprocal changes to their bylaws in order to establish
regular participation with the DefCore Committee as well.

This patch also adds a provision to the DefCore Committee procedures
requiring the DefCore Committee to produce a regular rollup of
prioritized issues to the Technical and User Committee.  The report is
intended to provide the Technical Committee with a prioritized list of
actionable interoperability issues and request the Technical Committee's
assistance and guidance in resolving them.  The report also serves as a
formal notification to the User Committee of interoperability challenges
that users may face, and request that the User Committee seek feedback
from the user community about how to prioritize and resolve them.

[1]
https://blueprints.launchpad.net/python-neutronclient/+spec/keystone-api-v3-support
[2] https://review.openstack.org/#/q/topic:bp/domain-scoped-tokens,n,z

[3]
http://git.openstack.org/cgit/openstack/governance/tree/reference/charter.rst#n10

[4] https://wiki.openstack.org/wiki/Governance/Foundation/UserCommittee

Change-Id: Ia3381b4dfb3d821ded38b15b6406f92a04d08c70