[RFC] Recurring testing recommendations
During the 2015-09-30 DefCore Committee meeting, some initial discussion
was held on the topic of how to handle recurring testing for vendor
products.  The Committee was asked by the Foundation's representative
for some guidance around how DefCore would like to handle recurring
testing in order to revise future legal agreements to be in sync with
DefCore.

This commit represents a first pass at a resolution for how to handle
recurring testing.  It is a strawman proposal intended to provoke
discussion between the DefCore Committee, Foundation staff, and other
interested parties.

Change-Id: I01daefbf4b69cef301809608a3f635647245ab3d