Require image operations
In recent weeks, the DefCore Committee [1][2][3][4], Board of Directors
[5], Technical Committee [6][7], as well as operators & developers
[8][9] within the community have discussed a number of questions related
to the level of interoperability that consumers of OpenStack Powered
products should be able to expect.  A rough consensus emerged that the
ability for end users to import their own images into a cloud and boot
instances from them emerged as an important factor for workload
interoperability.  In previous scoring discussions, we've scored several
capabilities related to image import and manipulation via the Glance v2
API favorably.

This patch moves those capabilities from advisory to required status in
the 2016.07 Guideline.

[1] https://review.openstack.org/#/c/244782/
[2] http://eavesdrop.openstack.org/meetings/defcore/2015/defcore.2015-11-18-16.00.html
[3] https://etherpad.openstack.org/p/DefCoreRing.DecWorkignSession
[4] http://eavesdrop.openstack.org/meetings/defore/2015/defore.2015-12-02-16.01.html
[5] https://wiki.openstack.org/wiki/Governance/Foundation/3Dec2015BoardMeeting
[6] http://eavesdrop.openstack.org/meetings/tc/2015/tc.2015-12-08-20.01.html
[7] http://lists.openstack.org/pipermail/openstack-tc/2015-December/001085.html
[8] http://lists.openstack.org/pipermail/openstack-operators/2015-November/008809.html
[9] http://lists.openstack.org/pipermail/openstack-dev/2015-November/079585.html

Change-Id: Ic5ebe66d06a5a27cce524c4383694fbba804e18f