Create rough draft of extension programs
For some time now, the Interop Working Group has been discussing the
possibility of creating new interoperability programs.  Specifically,
we're been considering programs for specific vertical use cases and
add-on programs for products that ship additional OpenStack software
outside what's required of the Powered program.  There have been some
questions in the community about what the design of such programs
might look like.  There are also ongoing discussions about related
topics like what tests should or should not be placed into the
Tempest tree (particularly with regards to these new programs).

Although we've discussed these programs at a number of meetings,
the PTG, and various other venues we still don't have a good
written description of what we're thinking and where we're at.

This patch attempts to lay out in writing what we in the Interop
Working Group have been thinking about the design of these programs.
The intent here is to provide a concrete document for discussion
so that the rest of the community can weigh in and not have to
speculate about these things as much.  It is not intended to be
a final document, but rather to provide some guidance to others
in the community as we continue to work on these definitions.

Change-Id: I8a44dcdf9d784134ae7f21dc11f7729e11d9a3c7